Add more dev dependencies
Vite is a fast build tool that offers hot reload for a better dev experience for frontend.

It also offers tree-shaking in which instead of importing a whole module that our code uses, only functions in that module that are used in the code will be imported.

Eslint and babel is also downloaded

This project will be written in typesript. With the type safety ts offers.we will be able to catch bugs earlier in development