Fix: protected route
Added new collab service call function to authenticate token before making requests to the backend. Alternative method is to add a "service" argument to the call function so that the call function would work with any backend service instead of the original question service hardcoded version, but time constraints and it would affect other code using the original callFunction.