[Collab] Hexagonal architecture, remove use of Redis (#62)
This PR mainly refactors existing skeleton code for collaboration
service.

`collaboration-service` now follows the hexagonal (ports and adapters)
architecture, similar to `user-service` and `question-service`.

This PR also removes the use of Redis as an intermediary for clients to
retrieve`userId-collabId` and other similar mappings. These will now be
retrieved directly from the `collaborationRepository`.

Lastly, this PR adds the supabase project-id for the collaboration table to the supabase-config.json file, for automatically generating types and enums.