[Collab] Document persistence, ws authentication middleware (#65)
## PR Description

This PR implements the functionality to persist the collaborative code
document into the Supabase `collaboration` table. The document shared
between two collaborators in a collaboration room will be fetched on
load, and updated every `x` seconds.

## Changes

**Frontend:**
1. Moved the collaboration page to `/collab/[id]`, where `id` serves as
the document name and identifier for the collaboration room.

**Backend:**

- **`collaboration-gateway`:**
1. Implemented functionality in the Hocuspocus server to fetch and save
the document to Supabase.
2. Added middleware for the WebSocket connection to ensure the below 3
criteria are met. If unfulfilled, connection will be closed.
      - The user is authenticated.
      - The collaboration entry exists in Supabase.
- The user is one of the collaborators in the specified collaboration
entry.
   3. Added logging for better traceability.
  
- **`collaboration-repository` and `collaboration-supabase`:**
   1. Added methods to support the above functionalities.