Add questionId field to add question
Originally, I had the questionId to be a Number field so there was no need for questionId. But since it is saved in the DB as a String, the comparisons did not work due to lexicographical issues. I was too lazy to implement a counter thus ill just request for a questionId when adding questions.